fix(sql): missing ApiLogger in WadoUriService
This commit updates the constructor of the WadoUriService class in the
retrieveInstance.js file. It now includes an instance of the ApiLogger
class as a parameter. Similarly, the constructor of the
SqlWadoUriService class in the WADO-URI.service.js file is also updated
to include the ApiLogger parameter. This change ensures that the logger
is properly passed to the service classes.